# src/main/resources/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Spring boot Postgres Docker Compose Api",
    "description": "Simple Spring Boot project based on PostgreSQL and Docker Compose",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/richart-proger/spring-boot-postgres-docker-compose-demo-2.git",
      "email": "arturgalimov33@gmail.com"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/students/item": {
      "put": {
        "summary": "Update student",
        "description": "Update student if it exists",
        "tags": [
          "student-rest-controller"
        ],
        "operationId": "update",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              },
              "examples": {
                "student" : {
                  "summary": "An example of a student",
                  "value":
                  {
                    "id" : 2,
                    "name" : "TOMAS",
                    "dob" : "2002-05-27",
                    "email" : "TOMAS@gmail.com"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "post": {
        "summary": "Create new student",
        "description": "Create new student",
        "tags": [
          "student-rest-controller"
        ],
        "operationId": "add",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              },
              "examples": {
                "student" : {
                  "summary": "An example of a student",
                  "value":
                  {
                    "name" : "Mary",
                    "dob" : "2002-02-27",
                    "email" : "mary@gmail.com"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "403": {
            "description": "Forbidden - Email constraint"
          }
        }
      }
    },
    "/api/students/list": {
      "get": {
        "tags": [
          "student-rest-controller"
        ],
        "summary": "Returns a list of students",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Student"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/students/item/{studentId}": {
      "delete": {
        "summary": "Delete student",
        "description": "Delete student by ID",
        "tags": [
          "student-rest-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "example": "1",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "dob": {
            "type": "string",
            "format": "date"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  }
}